Formalized processing packages as such. Made a location for scripts.
