+++ packages/processL0a/pyproject.toml
@@ -1,0 +1,9 @@
+[project]
+name = "processL0a"
+version = "0.1.0"
+description = "Handle raw data (Level 0a) from HAMMR-HD radiometer"
+requires-python = ">=3.12"
+dependencies = [
+    "tables",
+    "vitables[pyqt6]>=3.1.0", # Usage: vitables [h5filepath]
+]
