+++ scripts/makemeta.py
@@ -69,4 +69,3 @@ if __name__ == '__main__':
     filenames = filedialog.askopenfilenames()
     if filenames:
         make_metadata_file(filenames, context='cristalANT')
-# TODO loop the function not the argument
